+++ src/antity-pgsql.d.ts
@@ -72,6 +72,7 @@ export declare class SQLEntity extends Entity {
   private sel: unknown;
   private ins: unknown;
   private upd: unknown;
+  private arc: unknown;
   
   constructor(name: string, properties: Property[], schema?: string);
   
@@ -99,6 +100,15 @@ export declare class SQLEntity extends Entity {
     };
     
     update: (
+      rows: Record<string, unknown>[],
+      consumerId?: number | string,
+      consumerName?: string
+    ) => {
+      query: string;
+      args: unknown[];
+    };
+    
+    archive: (
       rows: Record<string, unknown>[],
       consumerId?: number | string,
       consumerName?: string